# [perf] Improve SchedulerMinHeap implementation