@acdlite 2021-04-06
Thanks for the PR! And for measuring the changes.

I think the original implementation used type checks to satisfy Flow (even though I forgot Flow intentionally pretends accessing an array of type `Array<T>` results in a `T`, since otherwise you have to add a bunch of unnecessary checks when doing loops), and then since it already checked the type, it didn't bother to also check the index. So, good catch!

I reverted the changes that weren't related to preventing out-of-bounds access, since they didn't change the algorithm. I don't really object to any of those changes, just in general I prefer not to merge stylistic changes since it introduces additional extra risk. Especially this code which is in an extremely hot path.

The `swap` function arguably aids readability but I prefer to keep that inlined so that we don't have to rely on Closure to inline it. (Even though Closure is gonna inline it anyway.)